Rename java package org.aerogear
