+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.jboss.aerogear</groupId>
-    <artifactId>jenkins-client</artifactId>
+    <artifactId>digger-java-client</artifactId>
     <version>1.0.0</version>
     <packaging>jar</packaging>
