+++ src/main/java/org/aerogear/digger/model/BuildStatus.java
@@ -1,4 +1,4 @@
-package com.redhat.digkins.model;
+package org.aerogear.digger.model;
 
 /**
  * Represents the status of a build.
