+++ src/main/java/org/aerogear/digger/util/DiggerClientException.java
@@ -1,4 +1,4 @@
-package com.redhat.digkins.util;
+package org.aerogear.digger.util;
 
 /**
  * Represents internal client exception
